Project Update: Cloud hop Clock Idea
This version contains a quick sketch in paint preview about what the clock will look like, as well as a thumbnail for the bl.ocks.org version of the little creature. I have also updated the README with the current plan for the project at this stage, which include extra things that I may do if I have time.

The next update I hope to have the p5.js files in the project folder, with some basic skeleton code.