Second class and 13th Handin Work
This is my version of the 3rd Maeda Clock. I have added the neccesary
files and updated all of the as well to the best of my knowledge. I
adapted my concept to fit inside the frame work of his design. The
Braille has been stretched and scaled around one point like in Maeda's
with the bottom of the braile heading towards the centre of the spiral.