Project Update: Maeda's Clock
This update contains a new version of the clock, based one off of John Maedas 12 clocks. I choose to remake the 5th clock. I also added a new feature where if you click on one of the numbers, the others disappear so you get a clearer look at the current hour, minute, or second. If you click anywhere on the screen the numbers come back.

I may add more to this over the next couple days.