Update to original Idea
Ive updated a lot of my code to the framwork that my final project will sit upon. The current ellipses are standins for image assets. Ive worked the code so that only once per complete rotation will the ellipses be wiped. This allows for an effect of a growing forest over a simple moving tree.

I also added a loop to each ring that postemptively draws all the trees as when starting the program, the trees would only be drawn from the current time, which causes an issue for any time that isnt exactly 12:00.