FINAL BUILD
Ive been working on this all night so I dont have incremental builds to push. The last build stop was my alarm build and from here I have been focusing on adding polish to the project.
I believe this point here is a good point to push the final build.