I decided to use images I’ve taken of Wellington CBD buildings at night. The images are black and white and I intend to keep them that way as I like the aesthetic.
I started experimenting with the code we were given.

values of each circle are taken in increments of 5 pixels.
elements that are masked change size depending on their brightness.

The mask that I’m using is just the image, as it is already black and white. White acts as the masked areas. Choosing to do this allows me more control over what areas the mask actually picks up, as I can change how bright the pixel has to be before it is part of the mask.

Nothing to fancy has been done yet. I am liking the simplicity and the contrast that has been produced but feel it needs something as it feels empty.