Ive changed a lot from the last version.
As a background I have created a grid but using x and y values that have been affected by noise. This produces a sort of random grid, but as noise functions different to the random function it doesn’t produce as extreme randomness.

In the masked areas, where brightness is above 128, pixels are represented by a small square.

The images produced are reminiscent of beads in a woven blanket.

I like the mesh like effect of the image. The grid brings a lot of dynamic to the image and can be looked at as a complex 3D object too. The squares (beads) are looking rather flat and I’d like to explore ways to develop them further.