The beads have been slightly updated. - The border has been removed - They change brightness based off of how bright the pixel they represent is.
The grid is remains the same as before.

Im not so sure I like how the beads are looking, in the last iteration they felt apart of the grid. Now, the image is disjointed, it is just a grid with squares onto and I’m not sure I like it much.