The beads have again been slightly updated. - They change size based off of how bright the pixel they represent is. - Opacity has been changed to 100 to give translucency
The grid is remains mostly the same but I changed the opacity of the lines to be slightly translucent and less bright/intense. This seems to help the contrast in the image as the image felt very dense in the last iteration.

Again I’m not happy as with the outcome as the two elements (grid and the squares) dont produce a cohesive image.