I ended up experimenting a lot to produce this outcome and I like the way it is looking. It produces a starry sky with constellations.
It took a lot of trial and error to arrive at this as I was trying to move away from my previous iteration as I felt stuck in its development.

The stars were created from a array where every 4th pixels x and y is recorded. The arrays order is then randomised.

The lines connect the first point of the mask to the last point. I don’t like the fact this produces a continuous line so will have to come up with some way to prevent this.

I like this idea of converting imagery into constellation design, but I would like to develop and abstract the idea further.