Commit Sketch 1
Each of my letters is composed with two rectangles and a triangle, which stands in relationship to their positioning. The size and position of the first rectangle is fixed, but the location and size of the second rectangle is controlled by three parameters. The triangle is tied to these first three parameters, but has its own parameters as well.

The three parameters per letter:
  * `RScale` : size of the smaller rectangle
  * `R1x` : x offset of the smaller rectangle relative to the first one
  * `R1y` : y offset of the smaller rectangle relative to the first one
  * `T1x` : x offset of the left triangle corner relative to the first one
  * `T1y` : y offset of the left triangle corner relative to the first one