Settled on a visual style
I was messing around with Bezier curves initally, trying to get them to form into letters, I was sucessful in creating an uppercase A, but realised at that point it was going to be too complicated to use these for the whole project. I've since moved to a much simpler style, using an ellipse and a rect to create minimal letter forms. I'll use these two shapes, overlapping and in various sizes, to create my alphabet.