Sketch.js
It took me a while but I think I finally got the hang of the assignment. It often takes me a while to get the hang of coding between each asssingments. I have finished my the three first letters. I decided to work with triangles , which is making it relatively complicated as I have to create many variables.