First attempt
Wasn't able to capture thumbnail img's as I believe that part of the code has an issue the tutors need to iron out. Also not entirely sure I'm doing the masking right because it doesn't seem to have an effect on my images. Will enquire about this. However I've been sucessful in getting three masks and input images into the code, as well as generating three differrent outputs with variations in the sketch.js code.