Updated masks and sketch.js code
Figured out that my mask images were far too large and that's why they weren't having an effect on the output. Fixed it by running the images through photoshop. So far I have a "painting effect" similar to what Phoebe created in one of the tutorial videos. The masks arent yet affecting it - will add this funciotnality soon. So far my inital idea for my theme that I quite like is "Nature in the city" (examining how and where we can escape to nature within Wellington city, and how that can be beneficial for all sorts of reasons, mental health etc.