Update README, and image files for addition of text, and background experimentation
For this commit, I explored the idea of abstracting my background further, so I stopped the animation. I also added text, to see if simple info in regards to the Kaka to see if it would add impact. I really liked the text, and so I plan to keep it. All files have been updated.