Final Push
I've got all three masks done and I'm really happy with the effects they create. It took some experimenting with the different layers of greyscale values to figure out what looks good where, with all the different effects going on. I've gotten to a point where I think they all look good, and I'm happy to hand them in now.

My theme is centred around the natural world, and appreciating the small, but valuable moments that arise when we take time to notice the things that my images depict. I really like the "painterly" effect my code generates in conjunction with the masks, I didn't want it to look too unnatural as that would be at odds with my theme. When I figured out getting the vertical strokes effect with ellipses, I think it works really well with the images I've used so I was glad I stumbled across it while messing around and experimenting. From there it was a matter of figuring out how I wanted to put emphasis on the subjects of the photographs, and I did this using masks created in Illustrator.

I hope that my theme comes across in the final output of my design, I feel that it's universally relatable because though a lot of us live in cities, we are still surrounded by nature, sometimes we're just not aware of it. I think it's incredibly valuable to take the time to stop once in a while and observe the natural world that's all around us, because it can ease your mind when you're stressed, and provide a bit of relief from the digital overload that we're usually faced with.

When I found out this project would be displayed on a billboard in the city, I soon came to the idea that it would be really nice for people to see a simple slice of the natural world in amongst the concrete jungle. In doing so I hope to remind people of how valuable nature is, and how connected we all are to the natural world. I strongly believe if our society was more connected with nature the world would be a much more pleasant place to live.

Overall I really enjoyed this project - one of my passions in life is photography so I jumped at the chance to have images that I've taken be used within a Creative Coding class, however I was pretty intimidated when I learned where they would be displayed. Gladly I'm really proud of the output of this project and I can't wait to go downtown and see them displayed!