My insperation for this project was one of my favourite art styles called Memphis. Memphis art is a vibrant style that was popular in the 1980s. it is known for its use of bright colours, geometric shapes and bold patterns.
I wanted to base my alphabet theme off Memphis style as I have always loved that kind of art. It seems really fun and lively and I thought it would be a cool concept for an alphabet.

I found that throughout this project there were a few challenges. My main challenge was making my alphabet legiable whilst also making it as abstact as possible. I didn't want to create a simple alphabet. However I also didn't want to make it so complext that it was impossible to figure out what each letter was. I am very happy with how each letter turned out as I feel they each represent my idea and theme perfectly.

For my shapes I designed the red and yellow objects so that they would revolve around the same point. This way it would give the sense all the shapes are connected and are part of a unit.

I have designed my default letter to resemble an exclamation point as I feel it best suits the bright and vibrant energy my work portays. All the words in the exhibition also reflect the fun and vibrant tone of my font.

For my animation I made the blue balls size change inbetween the interations of the letters and words. That way it gets really big then deflates into its next position. I decided to do this as it plays of the theme of exagerated fun and an explotions of colour.

I didn't want to do too many animations as it would be to much and it would seem like there is too much going on and you wouldn't know where to look.

overall I am very happy with my work. I have learnt a lot thought this project and I am proud of what I have accomplished.