Got everything working
The clock is working as intended, a fairly close replica of the Maeda clock I'm trying to recreate. My numbers use a segment display style, where every number is a subset of the active segments in the number 8. This doesn't read nearly as well as the Maeda version. To change this the coordinate values for each number could be changed, without the rest of the code needing any changes.