Transitions Between Times of Day
I decided to add in transitions in the sky colour between the hours of 4 and 5am, 8-9am, 5-6pm, and 8 - 9pm for the different parts of the day. It's not the best looking but it's only over a millisecond so at least it's there. I think going back to change it for over a second at this point is more complex than I need to get into.