Update clock.js
i have added an allarm clock that strobes  different colours. i have also added comments to help describe what is going on. i added a night time function. it changes the opacity of a rectangle for the background over a course of and hour and is black for the period of night time. it fades back out to red in the morning